refactor(core): split repair.rs into repair/ directory module (no behavior change)
Task 1 of the informed-consent adoption path (#374): mechanical split so
later tasks land in focused files. repair.rs (449 lines) becomes
repair/{mod,sweep,orchestration}.rs — sweep.rs holds
complete_pending_trash_renames, orchestration.rs holds repair_vault, and
mod.rs re-exports both so the public surface (crate::vault::repair_vault,
pub(crate) complete_pending_trash_renames) is unchanged. core/src/vault/mod.rs
needs no edit. Zero behavior change: verbatim code move plus import-path
adjustments only.
